+++ system/src/Grav/Common/Data/BlueprintSchema.php
@@ -54,7 +54,8 @@ class BlueprintSchema extends BlueprintSchemaBase implements ExportInterface
     public function validate(array $data)
     {
         try {
-            $messages = $this->validateArray($data, $this->nested, $this->items['']['form'] ?? []);
+            $validation = $this->items['']['form']['validation'] ?? 'loose';
+            $messages = $this->validateArray($data, $this->nested, $validation === 'strict');
         } catch (\RuntimeException $e) {
             throw (new ValidationException($e->getMessage(), $e->getCode(), $e))->setMessages();
         }
@@ -131,11 +132,11 @@ class BlueprintSchema extends BlueprintSchemaBase implements ExportInterface
     /**
      * @param array $data
      * @param array $rules
-     * @param array $parent
+     * @param bool $strict
      * @return array
      * @throws \RuntimeException
      */
-    protected function validateArray(array $data, array $rules, array $parent)
+    protected function validateArray(array $data, array $rules, bool $strict)
     {
         $messages = $this->checkRequired($data, $rules);
 
@@ -153,8 +154,8 @@ class BlueprintSchema extends BlueprintSchemaBase implements ExportInterface
                 $messages += Validation::validate($child, $rule);
             } elseif (\is_array($child) && \is_array($val)) {
                 // Array has been defined in blueprints.
-                $messages += $this->validateArray($child, $val, $rule ?? []);
-            } elseif (isset($parent['validation']) && $parent['validation'] === 'strict') {
+                $messages += $this->validateArray($child, $val, $strict);
+            } elseif ($strict) {
                 // Undefined/extra item.
                 throw new \RuntimeException(sprintf('%s is not defined in blueprints', $key));
             }