+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 1. [](#new)
     * Added `Session::regenerateId()` method to properly prevent session fixation issues
+1. [](#bugfix)
+    * Fixed `validation: strict` not working in blueprints [#1273](https://github.com/getgrav/grav/issues/1273)
 
 # v1.7.0-rc.6
 ## 02/11/2020